# docs/swagger/admin-api.yml
openapi: 3.1.0
info:
  title: Admin API
  description: Library of Coffee - Admin API
  contact:
    name: Library of Coffee
  version: 1.0.0
servers:
- url: /
  description: Current Server
security:
- Bearer: []
...
paths:
  /api/admin/shops/{id}:
    get:
      tags:
      - Shop
      summary: 店舗詳細取得
      description: 指定されたIDの店舗の詳細情報を取得します。
      operationId: getShop
      parameters:
      - name: id
        in: path
        description: 店舗ID
        required: true
        schema:
          type: string
        example: 00000000-0000-4000-8000-000000000001
      responses:
        "200":
          description: 取得成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShopDetailResponse"
              examples:
                success:
                  summary: 取得成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000001
                    shopifyShopId: test-shop-001
                    name: テスト珈琲店
                    introduction: こだわりの珈琲をお届けします。
                    particular: 厳選された豆のみを使用しています。
                    shopUrl: https://example.com
                    prefecture: TOKYO
                    publishStatus: PUBLISHED
                    images:
                    - id: 00000000-0000-4000-8000-000000000010
                      type: MAIN
                      imageUrl: https://example.com/shop-image.jpg
                    - id: 00000000-0000-4000-8000-000000000011
                      type: LOGO
                      imageUrl: https://example.com/shop-logo.png
        "404":
          description: 店舗が見つかりません
          content:
            '*/*':
              examples:
                not_found:
                  summary: 店舗が見つからない場合
                  description: not_found
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 404
                    error: Not Found
                    path: /api/admin/shops/00000000-0000-4000-8000-000000000999
    put:
      tags:
      - Shop
      summary: 店舗編集
      description: 指定されたIDの店舗を編集します。multipart/form-data形式で送信してください。全項目を置換します。ShopImageIdはサーバー側でUUIDv4を自動再生成します。
      operationId: updateShop
      parameters:
      - name: id
        in: path
        description: 店舗ID
        required: true
        schema:
          type: string
        example: 00000000-0000-4000-8000-000000000001
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/UpdateShopRequest"
                images:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
              - data
      responses:
        "200":
          description: 編集成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShopResponse"
              examples:
                success:
                  summary: 編集成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000001
        "400":
          description: バリデーションエラー
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad_request:
                  summary: バリデーションエラーの場合
                  description: bad_request
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 400
                    error: Bad Request
                    path: /api/admin/shops/00000000-0000-4000-8000-000000000001
        "404":
          description: 店舗が見つかりません
          content:
            '*/*':
              examples:
                not_found:
                  summary: 店舗が見つからない場合
                  description: not_found
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 404
                    error: Not Found
                    path: /api/admin/shops/00000000-0000-4000-8000-999999999999
        "409":
          description: ShopifyショップIDが重複しています
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                conflict:
                  summary: ShopifyショップIDが重複している場合
                  description: conflict
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 409
                    error: Conflict
                    path: /api/admin/shops/00000000-0000-4000-8000-000000000001
    delete:
      tags:
      - Shop
      summary: 店舗削除
      description: 指定されたIDの店舗を削除します。店舗画像も同時に削除されます。
      operationId: deleteShop
      parameters:
      - name: id
        in: path
        description: 店舗ID
        required: true
        schema:
          type: string
        example: 00000000-0000-4000-8000-000000000001
      responses:
        "204":
          description: 削除成功
        "404":
          description: 店舗が見つかりません
          content:
            '*/*':
              examples:
                not_found:
                  summary: 店舗が見つからない場合
                  description: not_found
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 404
                    error: Not Found
                    path: /api/admin/shops/00000000-0000-4000-8000-999999999999
        "409":
          description: 関連するコーヒー豆が存在するため削除できません
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                conflict:
                  summary: 関連するコーヒー豆が存在する場合
                  description: conflict
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 409
                    error: Conflict
                    path: /api/admin/shops/00000000-0000-4000-8000-000000000001
...
  /api/admin/shops:
    get:
      tags:
      - Shop
      summary: 店舗一覧取得
      description: 店舗の一覧をページネーション付きで取得します。店名による部分一致検索が可能です。画像は含まれません。
      operationId: listShops
      parameters:
      - name: page
        in: query
        description: ページ番号（0始まり）
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: size
        in: query
        description: 1ページあたりの件数
        required: false
        schema:
          type: integer
          format: int32
          default: 20
        example: 20
      - name: name
        in: query
        description: 店名（部分一致検索）
        required: false
        schema:
          type: string
        example: 珈琲
      responses:
        "200":
          description: 取得成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShopListResponse"
              examples:
                success:
                  summary: 取得成功例
                  description: success
                  value:
                    items:
                    - id: 00000000-0000-4000-8000-000000000001
                      shopifyShopId: test-shop-001
                      name: テスト珈琲店
                      introduction: こだわりの珈琲をお届けします。
                      particular: 厳選された豆のみを使用しています。
                      shopUrl: https://example.com
                      prefecture: TOKYO
                      participationStatus: PARTICIPATING
                    totalCount: 1
                    page: 0
                    size: 20
    post:
      tags:
      - Shop
      summary: 店舗登録
      description: 新しい店舗を登録します。multipart/form-data形式で送信してください。ShopIdおよびShopImageIdはサーバー側でUUIDv4を自動生成します。
      operationId: createShop
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/CreateShopRequest"
                images:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
              - data
      responses:
        "201":
          description: 登録成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ShopResponse"
              examples:
                success:
                  summary: 登録成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000001
        "400":
          description: バリデーションエラー
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad_request:
                  summary: バリデーションエラーの場合
                  description: bad_request
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 400
                    error: Bad Request
                    path: /api/admin/shops
        "409":
          description: ShopifyショップIDが重複しています
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                conflict:
                  summary: ShopifyショップIDが重複している場合
                  description: conflict
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 409
                    error: Conflict
                    path: /api/admin/shops
...
components:
  schemas:
    UpdateShopRequest:
      type: object
      description: 店舗編集リクエスト
      properties:
        shopifyShopId:
          type: string
          description: ShopifyのショップID
          example: test-shop-001
        name:
          type: string
          description: 店舗名
          example: テスト店舗
        introduction:
          type: string
          description: 店舗紹介
          example: テスト紹介文
        particular:
          type: string
          description: こだわり
          example: テストこだわり
        shopUrl:
          type: string
          description: 店舗URL
          example: https://example.com
        prefecture:
          type: string
          description: 都道府県
          example: TOKYO
        participationStatus:
          type: string
          description: "参画ステータス（BEFORE_PARTICIPATION: 参画前 / PARTICIPATING: 参画中 / DROPPED:\
            \ 参画落ち）"
          example: PARTICIPATING
        imageTypes:
          type: array
          description: アップロード画像のタイプ一覧（imagesと同じ順序）
          example:
          - MAIN
          items:
            type: string
        keepImageIds:
          type: array
          description: 保持する既存画像のID一覧（含まれない既存画像は削除される）
          items:
            type: string
      required:
      - name
      - participationStatus
      - prefecture
      - shopUrl
      - shopifyShopId
    ShopResponse:
      type: object
      description: 店舗登録レスポンス
      properties:
        id:
          type: string
          description: 店舗ID
          example: 00000000-0000-4000-8000-000000000001
      required:
      - id
    ErrorResponse:
      type: object
      description: エラーレスポンス
      properties:
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: 2026-02-23T12:00:00.000+00:00
        status:
          type: integer
          format: int32
          description: HTTPステータスコード
          example: 404
        error:
          type: string
          description: エラー概要
          example: Not Found
        message:
          type: string
          description: エラーメッセージ
          example: tastes must not be empty
        path:
          type: string
          description: リクエストパス
          example: /api/admin/resources/00000000-0000-4000-8000-000000000099
      required:
      - error
      - path
      - status
      - timestamp
...
    CreateShopRequest:
      type: object
      description: 店舗登録リクエスト
      properties:
        shopifyShopId:
          type: string
          description: ShopifyのショップID
          example: test-shop-001
        name:
          type: string
          description: 店舗名
          example: テスト店舗
        introduction:
          type: string
          description: 店舗紹介
          example: テスト紹介文
        particular:
          type: string
          description: こだわり
          example: テストこだわり
        shopUrl:
          type: string
          description: 店舗URL
          example: https://example.com
        prefecture:
          type: string
          description: 都道府県
          example: TOKYO
        participationStatus:
          type: string
          description: "参画ステータス（BEFORE_PARTICIPATION: 参画前 / PARTICIPATING: 参画中）"
          example: BEFORE_PARTICIPATION
        imageTypes:
          type: array
          description: アップロード画像のタイプ一覧（imagesと同じ順序）
          example:
          - MAIN
          items:
            type: string
      required:
      - name
      - participationStatus
      - prefecture
      - shopUrl
      - shopifyShopId
...
    ShopListResponse:
      type: object
      description: 店舗一覧レスポンス
      properties:
        items:
          type: array
          description: アイテム一覧
          items:
            $ref: "#/components/schemas/ShopSummaryResponse"
        totalCount:
          type: integer
          format: int64
          description: 全件数
          example: 42
        page:
          type: integer
          format: int32
          description: 現在のページ番号（0始まり）
          example: 0
        size:
          type: integer
          format: int32
          description: 1ページあたりの件数
          example: 20
      required:
      - items
      - page
      - size
      - totalCount
    ShopSummaryResponse:
      type: object
      description: 店舗一覧アイテム
      properties:
        id:
          type: string
          description: 店舗ID
          example: 00000000-0000-4000-8000-000000000001
        shopifyShopId:
          type: string
          description: ShopifyショップID
          example: test-shop-001
        name:
          type: string
          description: 店舗名
          example: テスト珈琲店
        introduction:
          type: string
          description: 紹介文
          example: こだわりの珈琲をお届けします。
        particular:
          type: string
          description: こだわり
          example: 厳選された豆のみを使用しています。
        shopUrl:
          type: string
          description: 店舗URL
          example: https://example.com
        prefecture:
          type: string
          description: 都道府県
          example: TOKYO
        participationStatus:
          type: string
          description: "参画ステータス（BEFORE_PARTICIPATION: 参画前 / PARTICIPATING: 参画中 / DROPPED:\
            \ 参画落ち）"
          example: PARTICIPATING
      required:
      - id
      - name
      - participationStatus
      - prefecture
      - shopUrl
      - shopifyShopId
    ImageDetail:
      type: object
      description: 画像詳細
      properties:
        id:
          type: string
          description: 画像ID
          example: 00000000-0000-4000-8000-000000000010
        type:
          type: string
          description: 画像種別
          example: MAIN
        imageUrl:
          type: string
          description: 画像URL
          example: https://example.com/image.jpg
      required:
      - id
      - imageUrl
      - type
    ShopDetailResponse:
      type: object
      description: 店舗詳細レスポンス
      properties:
        id:
          type: string
          description: 店舗ID
          example: 00000000-0000-4000-8000-000000000001
        shopifyShopId:
          type: string
          description: ShopifyショップID
          example: test-shop-001
        name:
          type: string
          description: 店舗名
          example: テスト珈琲店
        introduction:
          type: string
          description: 店舗紹介
          example: こだわりの珈琲をお届けします。
        particular:
          type: string
          description: こだわり
          example: 厳選された豆のみを使用しています。
        shopUrl:
          type: string
          description: 店舗URL
          example: https://example.com
        prefecture:
          type: string
          description: 都道府県
          example: TOKYO
        participationStatus:
          type: string
          description: "参画ステータス（BEFORE_PARTICIPATION: 参画前 / PARTICIPATING: 参画中 / DROPPED:\
            \ 参画落ち）"
          example: PARTICIPATING
        images:
          type: array
          description: 画像一覧
          items:
            $ref: "#/components/schemas/ImageDetail"
      required:
      - id
      - images
      - name
      - participationStatus
      - prefecture
      - shopUrl
      - shopifyShopId
...
  securitySchemes:
    Bearer:
      type: http
      description: JWTアクセストークン
      scheme: bearer
      bearerFormat: JWT